Added new Article and Game get methods, refactoring

+++ backend/src/main/java/com/gamebroadcast/forum/article/ArticleController.java
@@ -1,65 +1,94 @@
 package com.gamebroadcast.forum.article;
 
+import com.gamebroadcast.forum.article.models.ArticleAddUpdate;
+import com.gamebroadcast.forum.article.models.ArticleFullInfoVM;
+import com.gamebroadcast.forum.article.models.ArticleVM;
 import com.gamebroadcast.forum.exceptions.ApiRequestException;
-import org.springframework.beans.factory.annotation.Autowired;
+import com.gamebroadcast.forum.exceptions.NoEditRightsException;
+import com.gamebroadcast.forum.security.SessionUtils;
+
 import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
+
+import lombok.RequiredArgsConstructor;
 
 import java.util.List;
 
 @RestController
 @RequestMapping(path = "api/article")
+@RequiredArgsConstructor
 public class ArticleController {
 
     private final ArticleService articleService;
 
-    @Autowired
-    public ArticleController(ArticleService articleService) {
-        this.articleService = articleService;
-    }
-
     @GetMapping
-    public List<Article> getAllArticles() {
+    public List<ArticleVM> getAllArticles() {
         return articleService.getAllArticles();
     }
 
     @GetMapping(path = "/{articleId}")
-    public Article getArticle(@PathVariable("articleId") Long articleId) {
+    public ArticleVM getArticle(@PathVariable("articleId") Long articleId) {
         try {
-            return articleService.getArticle(articleId);
-        } catch (IllegalStateException e) {
+            return articleService.getArticleById(articleId);
+        } catch (RuntimeException e) {
+            throw new ApiRequestException(e.getMessage());
+        }
+    }
+
+    @GetMapping(path = "/FullInfo/{articleId}")
+    public ArticleFullInfoVM getArticleFullInfo(@PathVariable("articleId") Long articleId) {
+        try {
+            return articleService.getArticleFullInfoById(articleId);
+        } catch (RuntimeException e) {
             throw new ApiRequestException(e.getMessage());
         }
     }
 
     @PostMapping
     @ResponseStatus(value = HttpStatus.CREATED)
-    public void addArticle(@RequestBody Article newArticle) {
+    @PreAuthorize("hasRole('EDITOR')")
+    public void addArticle(@RequestBody ArticleAddUpdate newArticle) {
         try {
-            articleService.addArticle(newArticle);
-        } catch (IllegalStateException e) {
+            articleService.addArticle(newArticle, newArticle.content);
+        } catch (RuntimeException e) {
             throw new ApiRequestException(e.getMessage());
         }
     }
 
     @PutMapping(path = "/{articleId}")
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
-    public void updateArticle(@PathVariable("articleId") Long articleId, @RequestBody Article article) {
+    @PreAuthorize("hasRole('EDITOR')")
+    public void updateArticle(@PathVariable("articleId") Long articleId, @RequestBody ArticleAddUpdate articleUpdate) {
         try {
-            article.setId(articleId);
-            articleService.updateArticle(article);
-        } catch (IllegalStateException e) {
+            if (!sessionUserCanEditArticle(articleId)) {
+                throw new NoEditRightsException("article");
+            }
+            articleService.updateArticle(articleId, articleUpdate);
+        } catch (RuntimeException e) {
             throw new ApiRequestException(e.getMessage());
         }
     }
 
     @DeleteMapping(path = "/{articleId}")
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
+    @PreAuthorize("hasRole('EDITOR')")
     public void deleteArticle(@PathVariable("articleId") Long articleId) {
         try {
+            if (!sessionUserCanDeleteArticle(articleId)) {
+                throw new NoEditRightsException("article");
+            }
             articleService.deleteArticle(articleId);
-        } catch (IllegalStateException e) {
+        } catch (RuntimeException e) {
             throw new ApiRequestException(e.getMessage());
         }
+    }
+
+    private boolean sessionUserCanEditArticle(Long id) {
+        return articleService.sessionUserIsOwner(id);
+    }
+
+    private boolean sessionUserCanDeleteArticle(Long id) {
+        return articleService.sessionUserIsOwner(id) || SessionUtils.getUserFromSession().getRole().equals("ADMIN");
     }
 }
